Disable CURLOPT_FAILONERROR to properly set body for non 2xx-responses.
When making a request that returns in a status code other than 2xx, Curl would
output the response to STDOUT instead of part of the response in the `curl_exec`
call. This would cause the Response object returned from Client::makeRequest to
contain nothing in the body. By telling curl not to fail on error, the body is
left intact in the response from `curl_exec`, and the body is set as expected
in the Response object returned from Client::makeRequest.